added login page and authentication working correctly

+++ authentication_server/functions/GenerateSecretKey.js
@@ -3,8 +3,9 @@ const fs = require('fs')
 
 const generateKey=()=>{
     const SecertKey = "SECRET_KEY="+crypt.randomBytes(128).toString('hex')
-
-    fs.writeFileSync(__dirname.replace('functions','.env'),SecertKey)
+    if (!fs.existsSync(__dirname.replace('functions','.env'))){
+        fs.writeFileSync(__dirname.replace('functions','.env'),SecertKey)
+    }
 }
 
 module.exports = generateKey
